requirements/base.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.4
python-decouple==3.0
static3==0.6.1
coverage==4.0.3
django-extensions==1.6.1
django-test-without-migrations==0.4
djangorestframework==3.3.3
scikit-learn==0.17.1
numpy==1.10.4
requests==2.9.1
pytz
